requirements.txt
torchvision
transformers
numpy
Pillow
opencv-python
huggingface-hub
ultralytics
bitsandbytes
qwen-vl-utils
google-genai
requests
scipy
scikit-learn
insightface
